Capture: SetImmediate
Bug: 463867631
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/277921
Commit-Queue: Gregg Tavares <gman@chromium.org>

+++ src/dawn/native/Commands.h
@@ -391,8 +391,8 @@ struct SetImmediatesCmd {
     SetImmediatesCmd();
     ~SetImmediatesCmd();
 
-    uint64_t offset;
-    uint64_t size;
+    uint32_t offset;
+    uint32_t size;
 };
 
 struct SetIndexBufferCmd {
